refactor(ui): move NumberField + StepperTile into components/ui
Relocate the two generic gym-input primitives into the shared base-UI folder
(components/ui/) and export them from its index.ts barrel, matching the existing
convention. The barrel is now the self-maintaining inventory of base primitives —
so docs/skills can point at the folder instead of hand-listing components.
