+++ web/src/components/ui/StepperTile.tsx
