+++ web/src/components/ui/index.ts
@@ -4,3 +4,5 @@ export { default as SegmentedControl } from './SegmentedControl'
 export { default as EmptyState } from './EmptyState'
 export { default as SectionHeader } from './SectionHeader'
 export { default as PageHeader } from './PageHeader'
+export { default as NumberField } from './NumberField'
+export { default as StepperTile } from './StepperTile'
