+++ web/src/pages/GymModeWorkout.tsx
@@ -11,8 +11,8 @@ import { muscleColor, muscleColorBordered, EQUIPMENT_LABEL, muscleToBodySlugs } 
 import { useTheme } from '../hooks/useTheme'
 import { useWorkoutSession } from '../stores/workoutSession'
 import { workoutAPI } from '../services/api'
-import StepperTile from '../components/StepperTile'
-import NumberField from '../components/NumberField'
+import StepperTile from '../components/ui/StepperTile'
+import NumberField from '../components/ui/NumberField'
 import DiscardConfirm from '../components/DiscardConfirm'
 import { clampStep, clampValue } from '../utils/number'
 import { displayWeight, displayToLbs } from '../stores/settings'
